test_auth_app/requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
python-dotenv~=1.0.1
djangorestframework~=3.14.0